backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
license = { file = "../LICENSE.md" }
classifiers = [
    "License :: OSI Approved :: GNU Affero General Public License v3",
]
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.13",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "mypy>=1.16.1",
    "ruff>=0.12.1",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "B", "C4", "UP", "ARG", "SIM", "TCH", "TID", "Q"]
ignore = ["E501", "B008", "F401", "F841", "W293", "ARG001", "N815"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
